Raw config for the "complex_modifications" key.
This is not a fully specified `karabiner.json` configuration. This is only the snippet that is required for the described keyboard remappings. Check at the time of use exactly where in the config file to insert this. (Also, confirm snippet itself doesn't need minor adjustments as these things change over time...!)